Backend/requirements.txt
pdfplumber==0.11.4
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
ollama==0.4.7
